+++ src/main/java/cn/com/topit/gl/dao/VInventory.java
@@ -7,7 +7,7 @@ public class VInventory implements GenericPojo,java.io.Serializable {
 	private String materid;
 	private Integer type;
 	private Date updateTime;
-	private Integer quantities;
+	private Double quantities;
 	private String principal;
 	private String matername;
 	private String matetypeid;
@@ -63,12 +63,12 @@ public class VInventory implements GenericPojo,java.io.Serializable {
 	}
 
 
-	public Integer getQuantities() {
+	public Double getQuantities() {
 		return quantities;
 	}
 
 
-	public void setQuantities(Integer quantities) {
+	public void setQuantities(Double quantities) {
 		this.quantities = quantities;
 	}
